+++ modules/nlu/src/backend/election/ambiguous.ts
@@ -1,6 +1,6 @@
 import * as sdk from 'botpress/sdk'
 import _ from 'lodash'
-import { allInRange } from './math'
+import { allInRange, scaleConfidences } from './math'
 import { NONE_INTENT } from './typings'
 
 export function detectAmbiguity(input: sdk.IO.EventUnderstanding): sdk.IO.EventUnderstanding {
@@ -9,8 +9,7 @@ export function detectAmbiguity(input: sdk.IO.EventUnderstanding): sdk.IO.EventU
   }
 
   // +- 10% away from perfect median leads to ambiguity
-  const totalConfInTopN = _.sumBy(input.intents, i => i.confidence)
-  const preds = input.intents.map(i => ({ ...i, confidence: i.confidence / totalConfInTopN }))
+  const preds = scaleConfidences(input.intents)
   const perfectConfusion = 1 / preds.length
   const low = perfectConfusion - 0.1
   const up = perfectConfusion + 0.1