+++ src/bp/core/bots/bots-router.ts
@@ -3,6 +3,7 @@ import { HTTPServer } from 'core/app/server'
 import { BotService } from 'core/bots'
 import { ConfigProvider } from 'core/config'
 import { ConverseRouter, ConverseService } from 'core/converse'
+import { MediaServiceProvider } from 'core/media'
 import { disableForModule } from 'core/routers'
 import {
   AuthService,
@@ -13,11 +14,11 @@ import {
   checkBotVisibility
 } from 'core/security'
 import { NLUService } from 'core/services/nlu/nlu-service'
-import { HintsRouter, HintsService } from 'core/user-code'
 import { WorkspaceService } from 'core/users'
 import express, { Express, RequestHandler, Router } from 'express'
 import { AppLifecycle, AppLifecycleEvents } from 'lifecycle'
 import _ from 'lodash'
+import path from 'path'
 import { URL } from 'url'
 
 import { NLURouter } from '../routers/bots/nlu'
@@ -28,7 +29,6 @@ export class BotsRouter extends CustomRouter {
   private needPermissions: (operation: string, resource: string) => RequestHandler
   private checkMethodPermissions: (resource: string) => RequestHandler
   private nluRouter: NLURouter
-  private hintsRouter: HintsRouter
   private converseRouter: ConverseRouter
 
   constructor(
@@ -38,8 +38,8 @@ export class BotsRouter extends CustomRouter {
     private workspaceService: WorkspaceService,
     private nluService: NLUService,
     private converseService: ConverseService,
-    private hintsService: HintsService,
     private logger: Logger,
+    private mediaServiceProvider: MediaServiceProvider,
     private httpServer: HTTPServer
   ) {
     super('Bots', logger, Router({ mergeParams: true }))
@@ -50,7 +50,6 @@ export class BotsRouter extends CustomRouter {
 
     this.nluRouter = new NLURouter(this.logger, this.authService, this.workspaceService, this.nluService)
     this.converseRouter = new ConverseRouter(this.logger, this.converseService, this.authService, this.httpServer)
-    this.hintsRouter = new HintsRouter(this.logger, this.hintsService, this.authService, this.workspaceService)
   }
 
   async setupRoutes(app: express.Express) {
@@ -59,7 +58,27 @@ export class BotsRouter extends CustomRouter {
 
     this.router.use('/converse', this.converseRouter.router)
     this.router.use('/nlu', this.nluRouter.router)
-    this.router.use('/hints', this.hintsRouter.router)
+
+    this.router.get(
+      '/media/:filename',
+      this.asyncMiddleware(async (req, res) => {
+        const botId = req.params.botId
+        const type = path.extname(req.params.filename)
+
+        const mediaService = this.mediaServiceProvider.forBot(botId)
+        const contents = await mediaService.readFile(req.params.filename).catch(() => undefined)
+        if (!contents) {
+          return res.sendStatus(404)
+        }
+
+        // files are never overwritten because of the unique ID
+        // so we can set the header to cache the asset for 1 year
+        return res
+          .set({ 'Cache-Control': 'max-age=31556926' })
+          .type(type)
+          .send(contents)
+      })
+    )
 
     this.router.get(
       '/',