+++ src/bp/core/notifications/index.ts
@@ -1,2 +1,0 @@
-export * from './notifications-repository'
-export * from './notifications-service'