+++ src/bp/ui-studio/src/web/views/FlowBuilder/diagram/index.tsx
@@ -24,12 +24,13 @@ import {
   DiagramEngine,
   DiagramWidget,
   NodeModel,
-  PointModel
+  PointModel,
+  SelectingAction
 } from 'storm-react-diagrams'
 import {
   buildNewSkill,
   closeFlowNodeProps,
-  copyFlowNode,
+  copyFlowNodes,
   createFlow,
   createFlowNode,
   fetchFlows,
@@ -232,12 +233,14 @@ class Diagram extends Component<Props> {
     this.setState({ expandedNodes: getExpandedNodes() })
 
     ReactDOM.findDOMNode(this.diagramWidget).addEventListener('click', this.onDiagramClick)
+    ReactDOM.findDOMNode(this.diagramWidget).addEventListener('mousedown', this.onMouseDown)
     ReactDOM.findDOMNode(this.diagramWidget).addEventListener('dblclick', this.onDiagramDoubleClick)
     document.getElementById('diagramContainer').addEventListener('keydown', this.onKeyDown)
   }
 
   componentWillUnmount() {
     ReactDOM.findDOMNode(this.diagramWidget).removeEventListener('click', this.onDiagramClick)
+    ReactDOM.findDOMNode(this.diagramWidget).removeEventListener('mousedown', this.onMouseDown)
     ReactDOM.findDOMNode(this.diagramWidget).removeEventListener('dblclick', this.onDiagramDoubleClick)
     document.getElementById('diagramContainer').removeEventListener('keydown', this.onKeyDown)
   }
@@ -524,6 +527,22 @@ class Diagram extends Component<Props> {
     )
   }
 
+  onMouseDown = (event: MouseEvent) => {
+    // Pressing the "ctrl key" triggers multiple node selection
+    // Reimplemented the MouseDown handler from the Diagram Widget
+    // https://github.com/projectstorm/react-diagrams/blob/v5.2.1/src/widgets/DiagramWidget.tsx
+
+    const model = this.diagramWidget.getMouseElement(event)
+    if (model === null && event.ctrlKey) {
+      this.diagramEngine.clearRepaintEntities()
+      const relative = this.diagramEngine.getRelativePoint(event.clientX, event.clientY)
+      this.diagramWidget.startFiringAction(new SelectingAction(relative.x, relative.y))
+      this.diagramWidget.state.document.addEventListener('mousemove', this.diagramWidget.onMouseMove)
+      this.diagramWidget.state.document.addEventListener('mouseup', this.diagramWidget.onMouseUp)
+      event.stopPropagation()
+    }
+  }
+
   onDiagramClick = (event: MouseEvent) => {
     const selectedNode = this.manager.getSelectedNode() as BlockModel
     const currentNode = this.props.currentFlowNode
@@ -553,7 +572,13 @@ class Diagram extends Component<Props> {
     }
 
     if (selectedNode && (selectedNode.oldX !== selectedNode.x || selectedNode.oldY !== selectedNode.y)) {
-      this.props.updateFlowNode({ x: selectedNode.x, y: selectedNode.y })
+      const nodesToMove = []
+      for (const node of this.diagramEngine.getDiagramModel().getSelectedItems() as NodeModel[]) {
+        if (node.type === 'block') {
+          nodesToMove.push({ x: node.x, y: node.y, id: node.id })
+        }
+      }
+      this.props.updateFlowNode(nodesToMove)
       Object.assign(selectedNode, { oldX: selectedNode.x, oldY: selectedNode.y })
     }
 
@@ -606,7 +631,12 @@ class Diagram extends Component<Props> {
   }
 
   copySelectedElementToBuffer() {
-    this.props.copyFlowNode()
+    this.props.copyFlowNodes(
+      this.diagramEngine
+        .getDiagramModel()
+        .getSelectedItems()
+        .map(el => el.id)
+    )
     Toaster.create({
       className: 'recipe-toaster',
       position: Position.TOP_RIGHT
@@ -736,7 +766,7 @@ const mapStateToProps = (state: RootReducer) => ({
   currentFlow: getCurrentFlow(state),
   currentFlowNode: getCurrentFlowNode(state),
   currentDiagramAction: state.flows.currentDiagramAction,
-  canPasteNode: Boolean(state.flows.nodeInBuffer),
+  canPasteNode: Boolean(state.flows.buffer?.nodes),
   emulatorOpen: state.ui.emulatorOpen,
   debuggerEvent: state.flows.debuggerEvent,
   zoomLevel: state.ui.zoomLevel,
@@ -756,7 +786,7 @@ const mapDispatchToProps = {
   updateFlowNode,
   switchFlow,
   updateFlow,
-  copyFlowNode,
+  copyFlowNodes,
   pasteFlowNode,
   refreshFlowsLinks,
   insertNewSkillNode,